Move SimpleLogger lib into app source tree
