+++ composer.json
@@ -31,11 +31,11 @@
         "erusev/parsedown" : "1.6.0",
         "fguillot/json-rpc" : "1.2.1",
         "fguillot/picodb" : "1.0.17",
-        "fguillot/simpleLogger" : "1.0.1",
         "fguillot/simple-validator" : "1.0.1",
         "fguillot/simple-queue" : "1.0.1",
         "paragonie/random_compat": "2.0.11",
         "pimple/pimple" : "3.2.2",
+        "psr/log": "~1.0",
         "swiftmailer/swiftmailer" : "5.4.8",
         "symfony/console" : "3.4.2",
         "symfony/event-dispatcher" : "3.4.2",
